backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
packaging==25.0
pipenv==2025.0.2
platformdirs==4.3.7
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
pytz==2025.2
setuptools==80.3.1
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.30.0
